Move components into elements
