--- src/machinable/engine/__init__.py
@@ -1,1 +1,0 @@
-from machinable.engine.engine import Engine
