--- src/machinable/engine/dry_engine.py
@@ -1,6 +1,0 @@
-from machinable.engine.engine import Engine
-
-
-class DryEngine(Engine):
-    def _dispatch(self, execution: "Execution"):
-        pass
