+++ tests/test_end_to_end.py
@@ -5,7 +5,7 @@ from pydantic.errors import ExtraError
 
 
 def test_end_to_end_execution(tmp_path):
-    ml.Repository(
+    ml.Storage(
         "machinable.storage.filesystem_storage", {"directory": str(tmp_path)}
     ).connect()
     ml.Project("./tests/samples/project").connect()
